Support for size-distribution zero bin removal.
Upcoming netCDF files will have the legacy zero bin from
size-distribution data removed.  This includes from the raw data array
and the concentration array (e.g. ACDP and CCDP, A1DC and C1DC).

This detects new files by looking for a global attribute in the netCDF:

		:SizeDistributionZeroBin = "no" ;

Or perhaps that should be named 'NoSizeDistributionZeroBin' and then you
don't actually need to check the value, just check if the attr exists.

If the attribute does not exist, it will be assumed it is an old file.

I also intend to increase the version number of the netCDF file to 2.0:

		:ConventionsVersion = "2.0" ;